Initial submission of the Colibra project into Mercurial.
This submission contains the source code at revision 1 of the Colibra project. The project was initially hosted in Perforce, but I am trying out Mercurial, which seems to be a better system for a small project like this. I still want to have the history of the files in Mercurial, so I am submitting the changes based on build 1 of Colibra, and I will submit subsequent changes into the database.