Adding all source files according to build 2 of Colibra.
This submission adds the version of the files that were shipped as build 2 of Colibra. In build 2, I added the necessary changes to source projects and the solution to support the implmentation in C# and VB.NET. So there were some structural changes that are now reflected in the repository.