Added AlignmentList collection class.
After a lot of thought, I really want to keep dependencies in AutoCAD or Civil 3D code to a minimum accross the projects. When I first added the functionality to access the alignments from a document, I implemented it to return a collection of object ids. This adds a dependency on client code to reference acdbmgd.dll, and may tempt callers to open the object id directly, which is not why I am working on Colibra.

I changed the implementation of the Document object to remove the GetAlignmentIds() method, and exposed an Alignments property that returns an AlignmentList class that I added. The AlignmentList class is instantiated internally by passing the object id collection of all the alignments, and so far exposes a Count property that allows to determine the number of alignments in the drawing.

I also added a source file to test the AlignmentList collection, which will be used for further testing.

Other changes made are related to adding comments for public methods and properties, as well as some internal methods. I decided that it will be a good idea to comment on internal methods, even if it only serves to keep my thoughts straight.