Added support for transactions.
After a lot of thought, I decided to provide a Transaction wrapper to encapsulate the AutoCAD document transaction object. This will remove dependencies in AutoCAD specific code in client code. Hiding the transaction object inside the Document class works for certain situations, but it is hard to use when more control is needed.

I removed the methods BeginUpdate() and EndUpdate() from the document object and added a StartTransaction() method that checks if a transaction is active and creates a new one if there is not.

I updated all the unit tests that required an active transaction to use the new mechanism.