Implemented ObjectNodeProvider and ByNameObjectSelector.
I created a IObjectSelector interface to implement object selection classes. For testing, I will use ByNameObjectSelector, which allows to select existing objects in a drawing by name.

I implemented an ObjectNodeProvider class that returns a reference to the object id collection for a specified type of object. Currently, it only supports the Alignment node.