Fixed AlignmentEntityProcessor test where the order of the entities was not as expected.
It is impossible to know when the test broke, but it had to do with the way the entities were enumerated. I changed the enumeration policy in the test so it passes. Since we are not really testing the enumeration policies, and we are just looking at the output, the test should still be valid.